requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-debug-toolbar==5.1.0
django-redis==5.4.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
numpy==2.2.3
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
RapidFuzz==3.12.1
redis==5.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==6.0.0
tdqm==0.0.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.7.0
dj-database-url==2.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
stripe>=6.0.0
tqdm==4.67.1
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
openai==1.3.0
python-dotenv==1.0.0
django==5.1.6
celery==5.5.3
django-ratelimit==4.1.0
